refactor: improve LayerZeroAdapter payment validation and introduce msg.value handling
- Enhanced payment validation in LayerZeroAdapter to ensure sufficient msg.value for both LayerZero fees and forwarding.
- Introduced a new internal function, _validateMsgValue, to streamline msg.value checks across payment methods.
- Updated error handling for InsufficientMsgValue to accept uint256 for required and provided values.
- Modified integration and unit tests to reflect changes in msg.value handling, ensuring comprehensive coverage for various payment scenarios.

+++ packages/chain-bridge/src/interfaces/IBridgeAdapter.sol
@@ -26,7 +26,7 @@ interface IBridgeAdapter {
     //////////////////////////////////////////////////////////////*/
 
     /// @notice Thrown when insufficient msg.value is provided for the specified msgValue
-    error InsufficientMsgValue(uint128 required, uint256 provided);
+    error InsufficientMsgValue(uint256 required, uint256 provided);
 
     /// @notice Thrown when a chain is not supported
     error UnsupportedChain();
